requirements.txt
anthropic>=0.25.0
google-cloud-bigquery>=3.11.0
google-cloud-bigquery[pandas]>=3.11.0
pandas>=2.0.0
matplotlib>=3.7.0
PyYAML>=6.0
python-dotenv>=1.0.0
db-dtypes>=1.1.0
